Reorganizing recursive_cata so it can be part of the public API
